+++ sdk/identity/Azure.Identity/tests/samples/OtherCredentialSnippets.cs
@@ -8,16 +8,16 @@ namespace Azure.Identity.Tests.samples
 {
     public class OtherCredentialSnippets
     {
-        public void AzurePipelinesServiceConnectionCredential_Example()
+        public void AzurePipelinesCredential_Example()
         {
-            #region Snippet:AzurePipelinesServiceConnectionCredential_Example
+            #region Snippet:AzurePipelinesCredential_Example
             // Replace the following values with the actual values for the service connection.
             string clientId = "<service_connection_client_id>";
             string tenantId = "<service_connection_tenant_id>";
             string serviceConnectionId = "<service_connection_id>";
 
             // Construct the credential.
-            var credential = new AzurePipelinesServiceConnectionCredential(tenantId, clientId, serviceConnectionId);
+            var credential = new AzurePipelinesCredential(tenantId, clientId, serviceConnectionId);
 
             // Use the credential to authenticate with the Key Vault client.
             var client = new SecretClient(new Uri("https://keyvault-name.vault.azure.net/"), credential);